Move projectsparqlclient to graph-commons
