+++ triples-generator/src/main/scala/io/renku/triplesgenerator/events/consumers/membersync/EventHandler.scala
@@ -31,7 +31,7 @@ import io.renku.lock.syntax._
 import io.renku.http.client.GitLabClient
 import io.renku.lock.Lock
 import io.renku.triplesgenerator.TgLockDB.TsWriteLock
-import io.renku.triplesstore.SparqlQueryTimeRecorder
+import io.renku.triplesstore.{ProjectSparqlClient, SparqlQueryTimeRecorder}
 import org.typelevel.log4cats.Logger
 import tsmigrationrequest.migrations.reprovisioning.ReProvisioningStatus
